Extract metricsCard to global components
